Make a gradient background
I wanted it to be less boring.
The App.vue CSS is not 'scoped' so I can include css for the body in it,
even though the body is outside the App's template.

The background for the SVG is white, and I gave the logo some white
space around it.