This is intended to stop the selection from the click
But it doesn't. Does the v-stream not support the modifier, maybe?

In other news, reading about that leads to implications that the browser
treats rapid-click events differently, something about "event.detail >
1" indicating a double click. I wonder if that has anything to do with
the removal of dots.

Ren tried the app and they click REALLY FAST and there are like no dots
man.