Give up on animating the falling money
There has to be a way, but I haven't figured it out.
transform:translate doesn't work on the svg that represents the bill
It doesn't work on the path, probably because that already has an SVG
translation on it?
it does work on the text, and only within the bounds of the bill-svg.

We need to move the bill SVG as a whole, so an animation has to be
applied at that level. But the CSS transforms don't seem to have an
effect, implying that I can't use CSS keyframe animations.

There's an animation functionality inside SVGs, but that seems harder.
Will have to go there, I guess, if we're gonna make the bills rain down.