Fallback to talking to yourself
Now if the websocket ever fails, you'll continue talking to yourself
and only yourself.

There are 2 console errors that I don't think I can suppress
AND more importantly, it doesn't tell you that it couldn't connect.

Next is to change the "team name" to something that the error in the websocket can modify.