Fix: it updates the score board on message receipt
Before this fix, it was updating the scoreboard only when I pushed the
report button.

This has something to do with vue reactivity magic.
I noticed that the objects "Noticed;" inside the component had one
reactive-ified element, and it was the local one. The other team
member's elements were not reactive-ified.

I changed it to create a new object every time, instead of overwriting
the property for the incoming message every time. Now all elements of
each (new) object are reactive-ified

This is pretty garbage
At least very inefficient

Also i think it means the whole teamSystem is getting this treatment??
because I pass it in as props.

Yuck!

But how else would I do it?
Based on the interwebs, the only other way I know to pass this stuff in,
is as globals.
https://jasonwatmore.com/post/2019/04/02/vuejs-rxjs-communicating-between-components-with-observable-subject

So this trash fire is fine for now.