The types are a lie
OK <- It took me a long time to type that K because vscode thought it was the start of a chord

afaict, vue-chartjs declares Line a class, but implements it as ... not a class. just an object.
So I don't see how this could ever work. Fucking lies
asshole thing
I can use Line as a mixin and then it doesn't fail miserably (just has type errors_
) but the chart still doesn't seem to work.

I switched to this because it had more specific types

The types are a lie.